requirements.txt
tensorflow~=2.13.0
tensorflow-addons~=0.21.0
faiss-cpu~=1.7.4